LineBot/LineBot V1/requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.1
future==0.18.2
gunicorn==19.9.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.14.0
MarkupSafe==1.1.1
PyMySQL==0.10.1
python-decouple==3.3
requests==2.25.0
selenium==3.141.0
soupsieve==2.1
urllib3==1.26.2
Werkzeug==1.0.1
